Updated logic as per updated scenario
