+++ callautomation-outboundcalling/requirements.txt
